+++ sample-resource-plugin/src/main/java/org/opensearch/sample/SampleResourceExtension.java
@@ -13,6 +13,7 @@ package org.opensearch.sample;
 
 import java.util.Set;
 
+import org.opensearch.sample.client.ResourceSharingClientAccessor;
 import org.opensearch.security.spi.resources.ResourceProvider;
 import org.opensearch.security.spi.resources.ResourceSharingExtension;
 import org.opensearch.security.spi.resources.client.ResourceSharingClient;
@@ -24,8 +25,6 @@ import static org.opensearch.sample.utils.Constants.RESOURCE_INDEX_NAME;
  */
 public class SampleResourceExtension implements ResourceSharingExtension {
 
-    private ResourceSharingClient CLIENT;
-
     @Override
     public Set<ResourceProvider> getResourceProviders() {
         return Set.of(new ResourceProvider(SampleResource.class.getCanonicalName(), RESOURCE_INDEX_NAME));
@@ -33,14 +32,6 @@ public class SampleResourceExtension implements ResourceSharingExtension {
 
     @Override
     public void assignResourceSharingClient(ResourceSharingClient resourceSharingClient) {
-        CLIENT = resourceSharingClient;
-    }
-
-    /**
-     * Returns the assigned resource sharing client
-     */
-    @Override
-    public ResourceSharingClient getResourceSharingClient() {
-        return this.CLIENT;
+        ResourceSharingClientAccessor.getInstance().setResourceSharingClient(resourceSharingClient);
     }
 }